This commit is the first phase of a repository restructuring. The code in its current state is broken. Future commits will fix it in the second phase.
_ ROOT/config directory contains the current config files.  These files
should be split into a configuration part that should remain, and an
install part that should be common to all these config files and that
should be moved to a single file in ROOT/install.
_ ROOT/install directory is currently empty, but should contain one (or
several) file(s) that downloads the external dependecies, install them
and compile the GPTune c codes.
_ ROOT/travis directory contains the travis files
_ Each directory in ROOT/example/* is related to a specific application
where all the related files should be.  The run* files have been moved
to their related directories, but the run* files executing several
applications have not been moved yet, as they need to be split first.
_ The former ROOT/example/postprocess directory has been split among the
specific applications directories
_ Each class within the ROOT/GPTune directory has been moved into a
separate file.